Attempt to simplify Code
The initial version will just clear the DGraph database on startup and
then batch insert the data using Transactions. In later versions I will
implement upserts.

+++ DGraphSample/DGraphSample/Csv/Model/Flight.cs
@@ -201,7 +201,7 @@ namespace DGraphSample.Csv.Model
         /// <summary>
         /// Distance between airports (miles).
         /// </summary>
-        public double? Distance { get; set; }
+        public float? Distance { get; set; }
 
         /// <summary>
         /// Distance Intervals, every 250 Miles, for Flight Segment.
